#196 [fix] : 캐시 파일 제거
